website: community people pics consistently spaced
Most of the bios have five lines of text, so that's the driver for the
layout except for @grubernaut's and @mbfrahry's where there's only three
lines. This makes the pictures following the short bios look misaligned.

This quick fix just always leaves enough room for five lines of bio text,
ensuring that the photos appear at a consistent vertical rhythm.

It's annoying to hard-code a particular value here, since this value won't
survive e.g. a change to the typesetting, but a more involved fix here
(using flexbox layout, or something else complicated) doesn't seem
warranted.